Move to test_dtypes.py
